+++ accessory/OccupancySensor.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const {Accessory} = require('../util/Accessory');
-const {addBinarySensorCharacteristic} = require('./characteristic/BinarySensor');
+const {addOccupancyDetectedCharacteristic} = require('./characteristic/BinarySensor');
 const {addBatteryWarningCharacteristic} = require('./characteristic/Battery');
 
 class OccupancySensorAccessory extends Accessory {
@@ -17,7 +17,7 @@ class OccupancySensorAccessory extends Accessory {
     _getPrimaryService() {
         this._log.debug(`Creating occupancy sensor service for ${this.name}`);
         let primaryService = new this.Service.OccupancySensor(this.name);
-        addBinarySensorCharacteristic.bind(this)(primaryService, this.Characteristic.OccupancyDetected);
+        addOccupancyDetectedCharacteristic.bind(this)(primaryService);
         addBatteryWarningCharacteristic.bind(this)(primaryService);
         return primaryService;
     }