+++ accessory/characteristic/CurrentTargetPosition.js
@@ -6,50 +6,78 @@ const {addCurrentStateCharacteristic, addTargetStateCharacteristic} = require('.
 const CURRENT_TARGET_POSITION_CONFIG = {
     item: "item",
     inverted: "inverted",
+    multiplier: "multiplier",
     stateItem: "stateItem",
-    stateItemInverted: "stateItemInverted"
+    stateItemInverted: "stateItemInverted",
+    stateItemMultiplier: "stateItemMultiplier",
+    manuMode: "manuMode"
 };
 
 function addCurrentPositionCharacteristic(service) {
-    let item, itemType, inverted;
+    let item, itemType, inverted, multiplier;
+    let manuMode = this._checkInvertedConf(CURRENT_TARGET_POSITION_CONFIG.manuMode);
     if(this._config[CURRENT_TARGET_POSITION_CONFIG.stateItem]) {
         [item, itemType] = this._getAndCheckItemType(CURRENT_TARGET_POSITION_CONFIG.stateItem, ['Rollershutter', 'Number', 'Switch', 'Contact']);
         inverted = this._checkInvertedConf(CURRENT_TARGET_POSITION_CONFIG.stateItemInverted);
+        multiplier = this._checkMultiplierConf(CURRENT_TARGET_POSITION_CONFIG.stateItemMultiplier, itemType);
     } else {
         [item, itemType] = this._getAndCheckItemType(CURRENT_TARGET_POSITION_CONFIG.item, ['Rollershutter', 'Number', 'Switch']);
         inverted = this._checkInvertedConf(CURRENT_TARGET_POSITION_CONFIG.inverted);
+        multiplier = this._checkMultiplierConf(CURRENT_TARGET_POSITION_CONFIG.multiplier, itemType);
     }
-    addCurrentStateCharacteristic.bind(this)(service, this.Characteristic.CurrentPosition, item, itemType, inverted, positionTransformation);
+
+    let targetCharacteristic = manuMode ? service.getCharacteristic(this.Characteristic.TargetPosition) : null;
+
+    addCurrentStateCharacteristic.bind(this)(service.getCharacteristic(this.Characteristic.CurrentPosition),
+        item,
+        itemType,
+        inverted,
+        positionTransformation.bind(this,
+            multiplier,
+            service.getCharacteristic(this.Characteristic.TargetPosition)
+        ),
+        targetCharacteristic
+    );
 }
 
 function addTargetPositionCharacteristic(service) {
     let [item, itemType] = this._getAndCheckItemType(CURRENT_TARGET_POSITION_CONFIG.item, ['Rollershutter', 'Number', 'Switch']);
     let inverted = this._checkInvertedConf(CURRENT_TARGET_POSITION_CONFIG.inverted);
-    let stateItem, stateItemType, stateItemInverted;
+    let multiplier = this._checkMultiplierConf(CURRENT_TARGET_POSITION_CONFIG.multiplier, itemType);
+    let stateItem, stateItemType, stateItemInverted, stateItemMultiplier;
 
     if(this._config[CURRENT_TARGET_POSITION_CONFIG.stateItem]) {
         [stateItem, stateItemType] = this._getAndCheckItemType(CURRENT_TARGET_POSITION_CONFIG.stateItem, ['Rollershutter', 'Number', 'Switch', 'Contact']);
         stateItemInverted = this._checkInvertedConf(CURRENT_TARGET_POSITION_CONFIG.stateItemInverted);
+        stateItemMultiplier = this._checkMultiplierConf(CURRENT_TARGET_POSITION_CONFIG.stateItemMultiplier, stateItemType);
     } else {
         stateItem = item;
         stateItemType = itemType;
         stateItemInverted = inverted;
+        stateItemMultiplier = multiplier;
     }
-    addTargetStateCharacteristic.bind(this)(service, this.Characteristic.TargetPosition, item, itemType, inverted, stateItem, stateItemType, stateItemInverted, positionTransformation);
+    // In order to know if this is the setter and apply `UP`/`DOWN` to it in case of 100 and 0
+    if(itemType === 'Rollershutter') {
+        itemType = `${itemType}Setter`;
+    }
+    addTargetStateCharacteristic.bind(this)(service.getCharacteristic(this.Characteristic.TargetPosition),
+        item,
+        itemType,
+        inverted,
+        stateItem,
+        stateItemType,
+        stateItemInverted,
+        positionTransformation.bind(this,
+            multiplier,
+            null
+        ),
+        positionTransformation.bind(this,
+            stateItemMultiplier,
+            null
+        )
+    );
 }
 
-    // Not sure about this:
-    //
-    // this._subscribeCharacteristic(service,
-    //     this.Characteristic.TargetPosition,
-    //     thisItem,
-    //     this._transformation.bind(this,
-    //         thisItemType,
-    //         thisInverted
-    //     )
-    // );
-
-// Todo: Maybe a clean fix for the problem above with subscribing to current state & Manu-Mode option
 function addPositionStateCharacteristic(service) {
     this._log.debug(`Creating position state characteristic for ${this.name}`);
 
@@ -63,7 +91,7 @@ function addHoldPositionCharacteristic(service) {
     try {
         let [item] = this._getAndCheckItemType(CURRENT_TARGET_POSITION_CONFIG.item, ['Rollershutter']);
 
-        this._log.debug(`Creating position state characteristic for ${this.name} with item ${item}`);
+        this._log.debug(`Creating hold position state characteristic for ${this.name} with item ${item}`);
 
         service.getCharacteristic(this.Characteristic.HoldPosition) // Never tested, since I don't know how to invoke it
             .on('set', setState.bind(this,
@@ -77,8 +105,7 @@ function addHoldPositionCharacteristic(service) {
     }
 }
 
-// Todo: Maybe some grace area, if target and acutall state differ a couple of percent?
-function positionTransformation(type, inverted, value) {
+function positionTransformation(multiplier, targetStateCharacteristic, type, inverted, value) {
     let transformedValue;
 
     let onCommand = type === 'Contact' ? "OPEN": "ON";
@@ -90,30 +117,56 @@ function positionTransformation(type, inverted, value) {
             if(value === onCommand) {
                 transformedValue = inverted ?
                     0 :
-                    100
+                    100;
             } else if (value === offCommand) {
                 transformedValue = inverted ?
                     100 :
-                    0
+                    0;
             } else {
                 if(value >= 50 && !(inverted)) {
-                    transformedValue = onCommand
+                    transformedValue = onCommand;
                 } else {
-                    transformedValue = offCommand
+                    transformedValue = offCommand;
                 }
             }
             break;
+        case 'RollershutterSetter':
         case 'Rollershutter':
         case 'Number':
-            if(inverted) {
-                transformedValue = 100 - value;
+            //This part is only invoked if this is used in a setter context and the item is a rollershutter
+            // Not sure if `inverted` should matter here
+            if(type === 'RollershutterSetter' && value === 100) {
+                transformedValue = `UP`;
+            } else if(type === 'RollershutterSetter' && value === 0) {
+                transformedValue = `DOWN`;
             } else {
-                transformedValue = value;
+
+                if (inverted) {
+                    transformedValue = Math.floor(100 - (parseFloat(value) * multiplier));
+                } else {
+                    transformedValue = Math.floor(parseFloat(value) * multiplier);
+                }
+                if (transformedValue >= 99) { // Weird not showing 100 or 0 bug of openHAB
+                    transformedValue = 100;
+                }
+                if (transformedValue <= 1) {
+                    transformedValue = 0;
+                }
+
+                const threshold = 3;
+                if (targetStateCharacteristic && targetStateCharacteristic.value !== transformedValue) {
+                    this._log.debug(`Checking if actual state is within threshold (${threshold}) of target state`);
+                    if ((targetStateCharacteristic.value > transformedValue && (targetStateCharacteristic.value - threshold) <= transformedValue) ||
+                        (targetStateCharacteristic.value < transformedValue && (targetStateCharacteristic.value + threshold) >= transformedValue)) {
+                        this._log.debug(`Actually assigning target state ${targetStateCharacteristic.value}, because its within the threshold (${threshold}) of the actual state ${transformedValue}`);
+                        transformedValue = targetStateCharacteristic.value;
+                    }
+                }
             }
             break;
     }
 
-    this._log.debug(`Transformed ${value} with inverted set to ${inverted} for ${this.name} to ${transformedValue}`);
+    this._log.debug(`Transformed ${value} with inverted set to ${inverted} and multiplier set to ${multiplier} for ${this.name} (type: ${type}) to ${transformedValue}`);
     return transformedValue;
 }