+++ accessory/characteristic/NumericSensor.js
@@ -6,22 +6,35 @@ const NUMERIC_CONFIG = {
     item: "item"
 };
 
-function addNumericSensorCharacteristic(service, characteristic) {
+function addNumericSensorCharacteristic(characteristic) {
     let [item] = this._getAndCheckItemType(NUMERIC_CONFIG.item, ['Number']);
 
     this._log.debug(`Creating numeric sensor characteristic for ${this.name} with ${item}`);
 
-    service.getCharacteristic(characteristic)
-        .on('get', getState.bind(this,
+    characteristic.on('get', getState.bind(this,
             item,
             parseFloat
         ));
 
-    this._subscribeCharacteristic(service,
-        characteristic,
+    this._subscribeCharacteristic(characteristic,
         item,
         parseFloat
     );
 }
 
-module.exports = {addNumericSensorCharacteristic};
+function addCurrentRelativeHumidityCharacteristic(service) {
+    addNumericSensorCharacteristic.bind(this)(service.getCharacteristic(this.Characteristic.CurrentRelativeHumidity));
+}
+
+function addCurrentAmbientLightLevelCharacteristic(service) {
+    addNumericSensorCharacteristic.bind(this)(service.getCharacteristic(this.Characteristic.CurrentAmbientLightLevel));
+}
+
+function addCurrentTemperatureCharacteristic(service) {
+    addNumericSensorCharacteristic.bind(this)(service.getCharacteristic(this.Characteristic.CurrentTemperature));
+}
+module.exports = {
+    addCurrentRelativeHumidityCharacteristic,
+    addCurrentAmbientLightLevelCharacteristic,
+    addCurrentTemperatureCharacteristic
+};