+++ util/Accessory.js
@@ -36,7 +36,7 @@ class Accessory {
         return this._services;
     }
 
-    _subscribeCharacteristic(service, characteristic, item, transformation) {
+    _subscribeCharacteristic(characteristic, item, transformation, callback) {
         this._log.debug(`Subscribing to changes for ${item}`);
         this._openHAB.subscribe(item, function(value, habItem) {
             if(value instanceof Error) {
@@ -45,7 +45,10 @@ class Accessory {
                 this._log(`Received push with new state for item ${habItem}: ${value}`);
                 let transformedValue = transformValue(transformation, value);
                 if(transformedValue !== null) {
-                    service.setCharacteristic(characteristic, transformedValue);
+                    characteristic.setValue(transformedValue, null, "openHABIgnore");
+                    if(callback && typeof(callback) === "function") {
+                        callback(transformedValue);
+                    }
                 }
             }
         }.bind(this));
@@ -86,6 +89,24 @@ class Accessory {
         } else {
             return false;
         }
+    }
+
+    _checkMultiplierConf(key, itemType) {
+        if(itemType === 'Number' || itemType === 'Rollershutter') {
+            if(this._config[key]) {
+                let parsedValue = parseFloat(this._config[key]);
+                if (!isNaN(parsedValue)) {
+                    return parsedValue;
+                } else {
+                    this._log.debug(`Not parsing multiplier for ${this.name}, because value (${this._config[key]}) is not parsable as float: Result ${parsedValue})`);
+                }
+            } else {
+                this._log.debug(`Not parsing multiplier for ${this.name}, because ${key} is not defined in config: ${JSON.stringify(this._config)}`);
+            }
+        } else {
+            this._log.debug(`Not parsing multiplier for ${this.name}, because of item's type ${itemType}`);
+        }
+        return 1;
     }
 
     _getAccessoryInformationService(modelDescription) {
@@ -141,28 +162,35 @@ function getState(habItem, transformation, callback) {
     }.bind(this));
 }
 
-function setState(habItem, transformation, state, callback) {
+// context and connectionID are variables giving information about the origin of the request. If a setValue/setCharacteristic is called, we are able to manipulate those.
+// If context is set to 'openHABIgnore' the actual set state will not be executed towards openHAB
+function setState(habItem, transformation, state, callback, context, connectionID) {
     let transformedState = transformValue(transformation, state);
     this._log.debug(`Change target state of ${this.name} [${habItem}] to ${state} (transformed to ${transformedState})`);
-    if(transformedState instanceof Error) {
-        this._log.error(transformedState.message);
-        if(callback && typeof callback === "function") {
-            callback(transformedState);
-        }
+    this._log.error(`Context: ${context}, connectionID: ${connectionID}`);
+    if(context === "openHABIgnore") {
+        callback();
     } else {
-        this._openHAB.sendCommand(habItem, `${transformedState}`, function (error) {
-            if (error) {
-                this._log.error(`Unable to send command: ${error.message}`);
-                if(callback && typeof callback === "function") {
-                    callback(error);
-                }
-            } else {
-                this._log(`Changed target state of ${this.name} [${habItem}] to ${transformedState}`);
-                if(callback && typeof callback === "function") {
-                    callback();
-                }
+        if(transformedState instanceof Error) {
+            this._log.error(transformedState.message);
+            if(callback && typeof callback === "function") {
+                callback(transformedState);
             }
-        }.bind(this));
+        } else {
+            this._openHAB.sendCommand(habItem, `${transformedState}`, function (error) {
+                if (error) {
+                    this._log.error(`Unable to send command: ${error.message}`);
+                    if(callback && typeof callback === "function") {
+                        callback(error);
+                    }
+                } else {
+                    this._log(`Changed target state of ${this.name} [${habItem}] to ${transformedState}`);
+                    if(callback && typeof callback === "function") {
+                        callback();
+                    }
+                }
+            }.bind(this));
+        }
     }
 }